Files created and a layout for start is created
New Files: index.html, index.js, and styles.css are created. On which,
the starting layout for the start of the game is implemented.
For now, the container for the start button along with the styling is
completed.
Next working on the layout for the Game window where option selection
and points along with the functionality will be added.